feat: add support to pixart-sigma & fix resolution binding cases
