+++ pipefuser/pipelines/__init__.py
@@ -1,7 +1,9 @@
 from .base_pipeline import PipeFuserPipelineBaseWrapper
-from .pixart_alpha import PipeFuserPixArtAlphaPipeline
+from .pipeline_pixart_alpha import PipeFuserPixArtAlphaPipeline
+from .pipeline_pixart_sigma import PipeFuserPixArtSigmaPipeline
 
 __all__ = [
     "PipeFuserPipelineBaseWrapper",
     "PipeFuserPixArtAlphaPipeline",
+    "PipeFuserPixArtSigmaPipeline",
 ]
